frd_v1/requirements.txt
numpy
torch
pandas
tqdm
pillow
scikit-learn
matplotlib
multiprocess
seaborn
SimpleITK
# Install PyRadiomics from GitHub main branch
git+https://github.com/Radiomics/pyradiomics.git@master#egg=pyradiomics